LayoutTests/http/tests/media/media-seeking-no-ranges-server-expected.txt


Test ended by:

Seek media file served by a server not supporting "Range" requests.
Verify that the 'ended' event fires upon finishing the playback.

EVENT(playing)
RUN(audio.currentTime = 0.5)

EVENT(ended)
END OF TEST
